Move sort proxy model into own file.
