+++ ultimarc/main.py
@@ -10,8 +10,9 @@ from PySide6 import QtCore, QtWidgets, QtQml
 from ultimarc import translate_gettext as _
 from ultimarc.tools import ToolContextManager
 from ultimarc.ui.device_class_model import DeviceClassModel
-from ultimarc.ui.devices_list_proxy_model import DeviceListProxyModel, DeviceListSortProxyModel
+from ultimarc.ui.devices_list_proxy_model import DeviceListProxyModel
 from ultimarc.ui.devices_model import DevicesModel
+from ultimarc.ui.devices_sort_proxy_model import DevicesSortProxyModel
 from ultimarc.ui.units import Units
 
 import ultimarc.qml.rc_assets
@@ -44,7 +45,7 @@ if __name__ == '__main__':
         # Local class instantiation
         units = Units()
         device_filter = DeviceListProxyModel()
-        device_sort = DeviceListSortProxyModel()
+        device_sort = DevicesSortProxyModel()
         devices = DevicesModel(args, tool_env)
         # Provide the list to the string model from the filter model
         device_class = DeviceClassModel()
