+++ ultimarc/ui/devices_list_proxy_model.py
@@ -11,21 +11,6 @@ from ultimarc.ui.devices_model import DeviceRoles
 _logger = logging.getLogger('ultimarc')
 
 
-class DeviceListSortProxyModel(QSortFilterProxyModel, QObject):
-    def __init__(self):
-        super().__init__()
-        self.sort(0)
-
-    def lessThan(self, source_left: QModelIndex, source_right: QModelIndex):
-        class_left = source_left.data(DeviceRoles.DEVICE_CLASS)
-        class_right = source_right.data(DeviceRoles.DEVICE_CLASS)
-        connected_left = source_left.data(DeviceRoles.CONNECTED)
-        connected_right = source_right.data(DeviceRoles.CONNECTED)
-        if connected_right == connected_left:
-            return class_left < class_right
-        return connected_right < connected_left
-
-
 class DeviceListProxyModel(QSortFilterProxyModel, QObject):
     _changed_front_page_ = Signal(bool)
     _changed_filter_class_ = Signal(str)
@@ -61,7 +46,6 @@ class DeviceListProxyModel(QSortFilterProxyModel, QObject):
 
     def set_filter_class(self, new_filter):
         if self._filter_class_ != new_filter:
-            _logger.debug(f'set_filter_class: {new_filter}')
             self._filter_class_ = new_filter
             self._changed_filter_class_.emit(self._filter_class_)
             self.invalidateFilter()
